==> notebook/1.2_Pathway_annotation.Rmd <==
---
title: "pathway_annotation"
output:
  pdf_document: default
  html_document: default
date: "2025-06-18"
---
This script has not to be run in the for loop together with the 
1_Pathway_annotation.Rmd script

Because it reads all the excel files that have been created from the
previous script

Script designed in order to merge all the tables,
to evaluate statistically significant differentially
expressed pathways between time points

## Complete table
```{r pressure, echo=FALSE}
library(readxl)
library(dplyr)
library(writexl)

directory <- "results/1_Pathway_annotation/tables/KEGG_path"
excel_files <- list.files(path = directory, pattern = "\\.xls[x]?$", full.names = TRUE)
file_names <- tools::file_path_sans_ext(basename(excel_files))

dfs <- setNames(lapply(excel_files, read_excel), file_names)

dfs_renamed <- lapply(names(dfs), function(nm) {
  df <- dfs[[nm]]
  df %>%
    rename_with(~paste0(nm, "_", .), -Pathway)
})

merged_all <- Reduce(function(x, y) full_join(x, y, by = "Pathway"), dfs_renamed)

head(merged_all)
```

```{r pressure, echo=FALSE}
merged_all <- as.data.frame(merged_all)
```

```{r pressure, echo=FALSE}
# now we will reorder the columns, and after that we will do:
# - overview of functional pathways detected
# - temporal trends and differential expression of pathways across stages
# - notable pathways linked to fermentation

# substitute NA values to zeros
merged_all[is.na(merged_all)] <- 0
rownames(merged_all) <- merged_all$Pathway
merged_all$Pathway <- NULL
```

```{r pressure, echo=FALSE}
head(merged_all)
```

Save the table in an excel file
```{r pressure, echo=FALSE}
path <- "results/1_Pathway_annotation/tables/KEGG_path/final_table.xlsx"
write_xlsx(merged_all, path, col_names = TRUE, format_headers = TRUE)
```

Here you have to manually insert the conditions
## DESeq2
```{r pressure, echo=FALSE}
library(DESeq2)
# set the condition name
condition <- factor(c("T0", "T0", "T0", 
                      "T1", "T1", "T1", 
                      "T2", "T2", "T2", 
                      "T3", "T3", "T3", 
                      "T5", "T5", "T5"))

# transform all the values in the cells in integers
merged_all <- data.frame(lapply(merged_all, as.integer))
rownames(merged_all) <- merged_all$Pathways

dds <- DESeqDataSetFromMatrix(countData = merged_all,
                              colData = data.frame(condition),
                              design = ~ condition)

# Run DESeq2 analysis
dds <- DESeq(dds)

# Obtain results for differential expression
res <- results(dds)

# View summary of results
summary(res)

# Filter for significantly different pathways
significant_res <- res[which(res$padj < 0.05 & abs(res$log2FoldChange) > 1), ]

# View the significant pathways
head(significant_res)

significant_pathways <- rownames(significant_res)

# compare T1 vs T0
res_T1_vs_T0 <- results(dds, contrast=c("condition", "T1", "T0"))

# Compare T2 vs T1
res_T2_vs_T1 <- results(dds, contrast = c("condition", "T2", "T1"))

# Compare T3 vs T0
res_T3_vs_T2 <- results(dds, contrast = c("condition", "T3", "T2"))

# Compate T5 vs T3
res_T5_vs_T3 <- results(dds, contrast = c("condition", "T5", "T3"))

#### T0 VS T1 #####
# Filter significant pathways (padj < 0.05 and |log2FoldChange| > 1)
significant_pathways_T1_vs_T0 <- res_T1_vs_T0[which(res_T1_vs_T0$padj < 0.05 & abs(res_T1_vs_T0$log2FoldChange) > 1), ]

# Select significant pathways for heatmap visualization
significant_pathways_t10 <- rownames(significant_pathways_T1_vs_T0)

# we decide to normalize count for visualization, 
# because we see there is variability between replicates and so we want to reduce this variability
# in particular we use the function rlog() that transforms the count data to the log2 scale in a way 
# which minimizes differences between samples for rows with small counts,
# and which normalizes respect to library size.
# The rlog transformation produces a similar variance stabilizing effect.
rld <- rlog(dds, blind = FALSE)  # blind = FALSE keeps group structure
rld_mat <- assay(rld)
```

## Heatmap of significant pathways
# T0 vs T1
```{r pressure, echo=FALSE}
library(pheatmap)
pheatmap(rld_mat[significant_pathways_t10, c("T0R1", "T0R2", "T0R3", "T1R1", "T1R2", "T1R3")], 
         cluster_rows = TRUE, 
         cluster_cols = FALSE, 
         show_rownames = TRUE, 
         show_colnames = TRUE,
         scale = "row",  # Standardize pathways
         color = colorRampPalette(c("blue", "white", "red"))(50),
         main="Pathway difference T0 vs T1")
```

## T1 vs T2
# Filter significant pathways (padj < 0.05 and |log2FoldChange| > 1)
```{r pressure, echo=FALSE}
significant_pathways_T2_vs_T1 <- res_T2_vs_T1[which(res_T2_vs_T1$padj < 0.05 & abs(res_T2_vs_T1$log2FoldChange) > 1), ]

# Select significant pathways for heatmap visualization
significant_pathways_t12 <- rownames(significant_pathways_T2_vs_T1)

# Create a heatmap of significant pathways
library(pheatmap)
pheatmap(rld_mat[significant_pathways_t12, c("T1R1", "T1R2", "T1R3", "T2R1", "T2R2", "T2R3")], 
         cluster_rows = TRUE, 
         cluster_cols = FALSE, 
         show_rownames = TRUE, 
         show_colnames = TRUE,
         scale = "row",  # Standardize pathways
         color = colorRampPalette(c("blue", "white", "red"))(50),
         main="Pathway difference between T1 vs T2")
```

## T2 vs T3
```{r pressure, echo=FALSE}
# no significant pathways for T2 vs T3
# Filter significant pathways (padj < 0.05 and |log2FoldChange| > 1)
significant_pathways_T2_vs_T3 <- res_T3_vs_T2[which(res_T3_vs_T2$padj < 0.05 & abs(res_T3_vs_T2$log2FoldChange) > 1), ]
# no significant pathways between T2 and T3

# Select significant pathways for heatmap visualization
significant_pathways_23 <- rownames(significant_pathways_T2_vs_T3)

# Create a heatmap of significant pathways
library(pheatmap)
pheatmap(rld_mat[significant_pathways, c("T2R1", "T2R2", "T2R3", "T3R1", "T3R2", "T3R3")], 
         cluster_rows = TRUE, 
         cluster_cols = FALSE, 
         show_rownames = TRUE, 
         show_colnames = TRUE,
         scale = "row",  # Standardize pathways
         color = colorRampPalette(c("blue", "white", "red"))(50),
         main="Pathway difference between T2 vs T3")
```

## T3 vs T5
```{r pressure, echo=FALSE}
# Filter significant pathways (padj < 0.05 and |log2FoldChange| > 1)
significant_pathways_T5_vs_T3 <- res_T5_vs_T3[which(res_T5_vs_T3$padj < 0.05 & abs(res_T5_vs_T3$log2FoldChange) > 1), ]

# Select significant pathways for heatmap visualization
significant_pathways_35 <- rownames(significant_pathways_T5_vs_T3)

# Create a heatmap of significant pathways
library(pheatmap)
pheatmap(rld_mat[significant_pathways_35, c("T3R1", "T3R2", "T3R3", "T5R1", "T5R2", "T5R3")], 
         cluster_rows = TRUE, 
         cluster_cols = FALSE, 
         show_rownames = TRUE, 
         show_colnames = TRUE,
         scale = "row",  # Standardize pathways
         color = colorRampPalette(c("blue", "white", "red"))(50),
         main = "Pathway difference between T3 VS T5")

# Take all the pathways that result significant from the differential analysis and
# See the temporal trend over time
signif_path_tot <- c(significant_pathways_t10, significant_pathways_t12, significant_pathways_23, significant_pathways_35)
head(signif_path_tot)

# Subset from the initial dataframe only that specific pathways
sub_merged <- merged_t5r3_new[rownames(merged_t5r3_new) %in% signif_path_tot, ]
dim(sub_merged)
```

```{r pressure, echo=FALSE}
sign_path_all <- as.data.frame(c(significant_pathways_t10, significant_pathways_t12, significant_pathways_23,
                                 significant_pathways_35))
```

```{r pressure, echo=FALSE}
write_xlsx(sign_path_all, "results/1_Pathway_annotation/tables/KEGG_path/all_significant_pathways.xlsx")
```

## Bubble plots
### BUBBLE PLOTS: analysis of potential pathways into different timings of fermentation ##########

```{r pressure, echo=FALSE}
# multiple bubble plots on one line
library(ggplot2)
facet_wrap(~Comparison, scales="free_x")

# before binding by rows all the tables, we add for each table the column "Comparison"
significant_pathways_T1_vs_T0$Comparison <- "T0 vs T1"
significant_pathways_T2_vs_T1$Comparison <- "T1 vs T2"
# significant_pathways_T2_vs_T3$Comparison <- "T2 vs T3"  # we have to delete this comparison since there are not significant different pathways
significant_pathways_T5_vs_T3$Comparison <- "T3 vs T5"

significant_pathways_T1_vs_T0$Pathways <- rownames(significant_pathways_T1_vs_T0)
significant_pathways_T2_vs_T1$Pathways <- rownames(significant_pathways_T2_vs_T1)
significant_pathways_T5_vs_T3$Pathways <- rownames(significant_pathways_T5_vs_T3)

all_comparisons <- rbind(
  significant_pathways_T1_vs_T0,
  significant_pathways_T2_vs_T1,
  significant_pathways_T5_vs_T3
)


all_comparisons$Comparison <- factor(
  all_comparisons$Comparison,
  levels = c("T0 vs T1", "T1 vs T2", "T3 vs T5")
)

all_comparisons$Pathways <- rownames(all_comparisons)

all_comparisons$direction <- ifelse(all_comparisons$log2FoldChange > 0, "latter", "former")

ggplot(all_comparisons, aes(x = baseMean, y = Pathways)) +
  geom_point(aes(size = -log10(pvalue), color = direction), alpha = 0.8) +
  scale_color_manual(values = c("former" = "orange", "latter" = "green"),
                     labels = c("former" = "Enriched in former", "latter" = "Enriched in latter")) +
  scale_size_continuous(
    range = c(2, 9.5),  # Adjust range to make bubbles visibly larger
    name = "-log10(p-value)"  # Optional: more informative legend
  ) +
  facet_wrap(~Comparison, scales = "free_x") +
  theme_minimal() +
  labs(
    title = "Pathway Activity Across Time Points",
    x = "Depth",
    y = "Pathway annotation",
    color = "Direction"
  ) + 
  theme(
    legend.position = "right",
    strip.text = element_text(face = "bold"),
    axis.text.y = element_text(size = 9)
  )
```

## TO vs T1
```{r pressure, echo=FALSE}
# convert the pathways from rownames to a new column 
significant_pathways_T1_vs_T0$Pathways <- rownames(significant_pathways_T1_vs_T0)
head(significant_pathways_T1_vs_T0)

significant_pathways_T1_vs_T0$log_p <- -log10(significant_pathways_T1_vs_T0$pvalue)

significant_pathways_T1_vs_T0$direction <- ifelse(significant_pathways_T1_vs_T0$log2FoldChange > 0, "latter", "former")

significant_pathways_T1_vs_T0$Comparison <- "T0 vs T1"

# Bubble plot
library(ggplot2)
ggplot(significant_pathways_T1_vs_T0, aes(x = baseMean, y = Pathways)) +
  geom_point(aes(size = log_p, color = direction), alpha = 0.8) +
  scale_color_manual(values = c("former" = "orange", "latter" = "green"),
                     labels = c("former" = "Enriched in former", "latter" = "Enriched in latter")) +
  facet_wrap(~Comparison, scales = "free_y") +
  theme_minimal() +
  labs(
    title = "T0 vs T1 Pathway Activity",
    x = "Depth",
    y = "Pathway annotation",
    size = "-log10(p-value)",
    color = "Direction"
  ) + 
  theme(
    legend.position = "right",
    strip.text = element_text(face = "bold"),
    axis.text.y = element_text(size = 9)
  )
```

Among these, consider only the significant pathways interesting for the alcoholic fermentation process and represent them in a histogram 
```{r pressure, echo=FALSE}
# recover the significant pathways that are related to alcoholic fermentation
significant_pathways_T1_vs_T0$Pathways

fermentation_pathways <- c(
  "ABC transporters", # linked to biofilm formation
  "Bacterial secretion system",
  "Biofilm formation - Pseudomonas aeruginosa",
  "Caffeine metabolism",
  "Carotenoid biosynthesis",
  "Cell cycle - Caulobacter",
  "Chlorocyclohexane and chlorobenzene degradation",
  "Fluorobenzoate degradation",
  "Quorum sensing",
  "Toluene degradation",
  "Lipopolysaccharide biosynthesis",
  "beta-Lactam resistance",
  "Dioxin degradation",
  "Nonribosomal peptide structures",
  "Peptidoglycan biosynthesis",
  "Phosphotransferase system (PTS)",
  "Vancomycin resistance",
  "Ethylbenzene degradation"
)

# take for this pathways the values os basemean in time t0 and t1 taking into account also the mean and standard deviation
sub_merged_all <- merged_all[which(rownames(merged_all) %in% fermentation_pathways),]
dim(sub_merged_all) # finally i have 6 pathways and 15 samples, for now we take only the ones at time T0 and T1
sub_merged_all <- sub_merged_all[,c("T0R1", "T0R2", "T0R3", "T1R1","T1R2", "T1R3")]
head(sub_merged_t0t1)
```

```{r pressure, echo=FALSE}
library(dplyr)
library(tibble)

df_summary <- sub_merged_t0t1 %>%
  tibble::rownames_to_column("Pathway") %>%   # 👈 save rownames into a "Pathway" column
  rowwise() %>%
  mutate(
    T0_mean = mean(c(T0R1, T0R2, T0R3)),
    T1_mean = mean(c(T1R1, T1R2, T1R3)),
    T0_sd   = sd(c(T0R1, T0R2, T0R3)),
    T1_sd   = sd(c(T1R1, T1R2, T1R3))
  ) %>%
  ungroup()

```

```{r pressure, echo=FALSE}
library(dplyr)
library(tidyr)

df_long <- df_summary %>%
  # bring rownames into a proper column
  select(Pathway, T0_mean, T1_mean, T0_sd, T1_sd) %>%
  pivot_longer(
    cols = starts_with("T"),
    names_to = c("Time", ".value"),
    names_pattern = "(T[01])_(.*)"
  )

```

```{r pressure, echo=FALSE}
library(ggplot2)
library(dplyr)

# now we log transform the values in the table beacuse otherwise 
# we cannot see the difference in values between time points 
# since the values are very different

# Add log10 transformed mean and approximate SE for error bars
df_summary <- df_long %>%
  mutate(
    log_mean = log10(mean + 1),       # add 1 to avoid log(0)
    # Approximate SE on log scale using delta method:
    log_se = sd / (mean * log(10) + 1e-8)  # adding small number to avoid div by zero
  )

png("results/1_Pathway_annotation/figures/bubble_plot_t0t1.png", width = 1200, height = 900, res = 200)  # 300 dpi is print-quality

ggplot(df_summary, aes(x = Pathway, y = log_mean, fill = Time)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.8)) +
  geom_errorbar(aes(ymin = log_mean - log_se, ymax = log_mean + log_se),
                width = 0.2,
                position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = c("T0" = "orange", "T1" = "green")) +  # Set colors here
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 11)  # increase size here
  ) +
  labs(
    y = "Log10 Mean Expression (+1)",
    x = "Pathway",
    title = "Pathway expression at T0 vs T1 (log10 scale)",
    fill = "Time"
  )

dev.off()
```

## T1 VS T2
```{r pressure, echo=FALSE}
significant_pathways_T2_vs_T1$Pathways <- rownames(significant_pathways_T2_vs_T1)
head(significant_pathways_T2_vs_T1)

significant_pathways_T2_vs_T1$log_p <- -log10(significant_pathways_T2_vs_T1$pvalue)

significant_pathways_T2_vs_T1$direction <- ifelse(significant_pathways_T2_vs_T1$log2FoldChange > 0, "latter", "former")

significant_pathways_T2_vs_T1$Comparison <- "T1 vs T2"

png("results/1_Pathway_annotation/figures/bubble_plot_t1t2.png", width = 1200, height = 900, res = 200) 

ggplot(significant_pathways_T2_vs_T1, aes(x = baseMean, y = Pathways)) +
  geom_point(aes(size = log_p, color = direction), alpha = 0.8) +
  scale_color_manual(values = c("former" = "orange", "latter" = "green"),
                     labels = c("former" = "Enriched in former", "latter" = "Enriched in latter")) +
  theme_minimal() +
  facet_wrap(~Comparison, scales = "free_y") +
  labs(
    title = "T1 vs T2 Pathway Activity",
    x = "Depth",
    y = "Pathway annotation",
    size = "-log10(p-value)",
    color = "Direction"
  ) + 
  theme(
    legend.position = "right",
    strip.text = element_text(face = "bold"),
    axis.text.y = element_text(size = 9)
  )

dev.off()
```
Among these, consider only the significant pathways interesting for the alcoholic fermentation process and represent them in a histogram 
```{r pressure, echo=FALSE}
# recover the significant pathways that are related to alcoholic fermentation
significant_pathways_T2_vs_T1$Pathways

fermentation_pathways <- c(
  "Bacterial secretion system",
  "ABC transporters",
  "Polycyclic aromatic hydrocarbon degradation",
  "Naphthalene degradation",
  "Furfural degradation",
  "Ethylbenzene degradation",
  "Dioxin degradation",
  "Cell cycle - Caulobacter",
  "Biofilm formation - Pseudomonas aeruginosa",
  "Bacterial secretion system",
  "ABC transporters"
)

# take for this pathways the values os basemean in time t0 and t1 taking into account also the mean and standard deviation
sub_merged_t1t2 <- merged_t5r3_new[which(rownames(merged_t5r3_new) %in% fermentation_pathways),]
dim(sub_merged_t1t2) # finally i have 6 pathways and 15 samples, for now we take only the ones at time T0 and T1
sub_merged_t1t2 <- sub_merged_t1t2[,c("T1R1", "T1R2", "T1R3", "T2R1","T2R2", "T2R3")]
head(sub_merged_t1t2)
```

```{r pressure, echo=FALSE}
library(dplyr)

df_summary <- sub_merged_t1t2 %>%
  rowwise() %>%
  mutate(
    T1_mean = mean(c(T1R1, T1R2, T1R3)),
    T2_mean = mean(c(T2R1, T2R2, T2R3)),
    T1_sd   = sd(c(T1R1, T1R2, T1R3)),
    T2_sd   = sd(c(T2R1, T2R2, T2R3))
  ) %>%
  ungroup()
df_summary$Pathway <- rownames(sub_merged_t1t2)
```

```{r pressure, echo=FALSE}
library(dplyr)
library(tidyr)

df_long <- df_summary %>%
  # bring rownames into a proper column
  select(Pathway, T1_mean, T2_mean, T1_sd, T2_sd) %>%
  pivot_longer(
    cols = starts_with("T"),
    names_to = c("Time", ".value"),
    names_pattern = "(T[12])_(.*)"
  )

```

```{r pressure, echo=FALSE}
library(ggplot2)
library(dplyr)

# now we log transform the values in the table beacuse otherwise 
# we cannot see the difference in values between time points 
# since the values are very different

# Add log10 transformed mean and approximate SE for error bars
df_summary <- df_long %>%
  mutate(
    log_mean = log10(mean + 1),       # add 1 to avoid log(0)
    # Approximate SE on log scale using delta method:
    log_se = sd / (mean * log(10) + 1e-8)  # adding small number to avoid div by zero
  )

png("results/1_Pathway_annotation/figures/bubble_plot_t1t2.png", width = 1200, height = 900, res = 200) 

ggplot(df_summary, aes(x = Pathway, y = log_mean, fill = Time)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.8)) +
  geom_errorbar(aes(ymin = log_mean - log_se, ymax = log_mean + log_se),
                width = 0.2,
                position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = c("T1" = "orange", "T2" = "green")) +  # Set colors here
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 13)  # increase size here
  ) +
  labs(
    y = "Log10 Mean Expression (+1)",
    x = "Pathway",
    title = "Pathway expression at T1 vs T2 (log10 scale)",
    fill = "Time"
  )

```

##### T3 vs T5 #####
```{r pressure, echo=FALSE}
significant_pathways_T5_vs_T3$Pathways <- rownames(significant_pathways_T5_vs_T3)
head(significant_pathways_T5_vs_T3)

significant_pathways_T5_vs_T3$log_p <- -log10(significant_pathways_T5_vs_T3$pvalue)

significant_pathways_T5_vs_T3$direction <- ifelse(significant_pathways_T5_vs_T3$log2FoldChange > 0, "latter", "former")

significant_pathways_T5_vs_T3$Comparison <- "T3 vs T5"

png("results/1_Pathway_annotation/figures/bubble_plot_t3t5.png", width = 1200, height = 900, res = 200) 

ggplot(significant_pathways_T5_vs_T3, aes(x = baseMean, y = Pathways)) +
  geom_point(aes(size = log_p, color = direction), alpha = 0.8) +
  scale_color_manual(values = c("former" = "orange", "latter" = "green"),
                     labels = c("former" = "Enriched in former", "latter" = "Enriched in latter")) +
  facet_wrap(~Comparison, scales = "free_y") +
  theme_minimal() +
  labs(
    title = "T3 vs T5 Pathway Activity",
    x = "Depth",
    y = "Pathway annotation",
    size = "-log10(p-value)",
    color = "Direction"
  ) + 
  theme(
    legend.position = "right",
    strip.text = element_text(face = "bold"),
    axis.text.y = element_text(size = 9)
  )

dev.off
```
Among these, consider only the significant pathways interesting for the alcoholic fermentation process and represent them in a histogram 
```{r pressure, echo=FALSE}
# recover the significant pathways that are related to alcoholic fermentation
significant_pathways_T5_vs_T3$Pathways

fermentation_pathways <- c(
  "Biofilm formation - Pseudomonas aeruginosa",
  "Caffeine metabolism",
  "Cell cycle - Caulobacter",
  "Oxidative phosphorylation",
  "Two-component system",
  "Penicillin and cephalosporin biosynthesis",
  "Ethylbenzene degradation",
  "Furfural degradation"
)

# take for this pathways the values os basemean in time t0 and t1 taking into account also the mean and standard deviation
sub_merged_t3t5 <- merged_t5r3_new[which(rownames(merged_t5r3_new) %in% fermentation_pathways),]
dim(sub_merged_t3t5) # finally i have 6 pathways and 15 samples, for now we take only the ones at time T0 and T1
sub_merged_t3t5 <- sub_merged_t3t5[,c("T3R1", "T3R2", "T3R3", "T5R1","T5R2", "T5R3")]
head(sub_merged_t3t5)
```

```{r pressure, echo=FALSE}
library(dplyr)

df_summary <- sub_merged_t3t5 %>%
  rowwise() %>%
  mutate(
    T3_mean = mean(c(T3R1, T3R2, T3R3)),
    T5_mean = mean(c(T5R1, T5R2, T5R3)),
    T3_sd   = sd(c(T3R1, T3R2, T3R3)),
    T5_sd   = sd(c(T5R1, T5R2, T5R3))
  ) %>%
  ungroup()
df_summary$Pathway <- rownames(sub_merged_t3t5)
```

```{r pressure, echo=FALSE}
library(dplyr)
library(tidyr)

df_long <- df_summary %>%
  # bring rownames into a proper column
  select(Pathway, T3_mean, T5_mean, T3_sd, T5_sd) %>%
  pivot_longer(
    cols = starts_with("T"),
    names_to = c("Time", ".value"),
    names_pattern = "(T[35])_(.*)"
  )

```

```{r pressure, echo=FALSE}
library(ggplot2)
library(dplyr)

# now we log transform the values in the table beacuse otherwise 
# we cannot see the difference in values between time points 
# since the values are very different

# Add log10 transformed mean and approximate SE for error bars
df_summary <- df_long %>%
  mutate(
    log_mean = log10(mean + 1),       # add 1 to avoid log(0)
    # Approximate SE on log scale using delta method:
    log_se = sd / (mean * log(10) + 1e-8)  # adding small number to avoid div by zero
  )

ggplot(df_summary, aes(x = Pathway, y = log_mean, fill = Time)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.8)) +
  geom_errorbar(aes(ymin = log_mean - log_se, ymax = log_mean + log_se),
                width = 0.2,
                position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = c("T3" = "orange", "T5" = "green")) +  # Set colors here
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 13)  # increase size here
  ) +
  labs(
    y = "Log10 Mean Expression (+1)",
    x = "Pathway",
    title = "Pathway expression at T3 vs T5 (log10 scale)",
    fill = "Time"
  )
```

Creation of a file with all the pathways that are significant and interesting (those that we plotted in the histograms)
```{r pressure, echo=FALSE}
pathways <- c(rownames(sub_merged_t0t1), rownames(sub_merged_t1t2), rownames(sub_merged_t3t5))
length(pathways)
unique_pathways <- unique(pathways)
length(unique_pathways)
```

Save in a file the unique pathways
```{r pressure, echo=FALSE}
df_unique_path <- as.data.frame(unique_pathways)
write_xlsx(df_unique_path, "results/1_Pathway_annotation/figures/unique_pathways.xlsx")
```
